Refactor ActiveCompensationRequestView
Make parts more modular to be able to reuse the compensation request
list and display better
