+++ gui/src/main/java/io/bisq/gui/main/dao/compensation/CompensationRequestDisplay.java
@@ -22,16 +22,19 @@ import io.bisq.core.btc.wallet.BsqWalletService;
 import io.bisq.core.dao.compensation.CompensationRequestPayload;
 import io.bisq.core.dao.compensation.Restrictions;
 import io.bisq.core.provider.fee.FeeService;
-import io.bisq.gui.components.HyperlinkWithIcon;
-import io.bisq.gui.components.InputTextField;
-import io.bisq.gui.components.TxIdTextField;
+import io.bisq.gui.components.*;
 import io.bisq.gui.util.BsqFormatter;
 import io.bisq.gui.util.GUIUtil;
 import io.bisq.gui.util.Layout;
 import io.bisq.gui.util.validation.BsqAddressValidator;
 import io.bisq.gui.util.validation.BsqValidator;
-import javafx.scene.control.TextArea;
-import javafx.scene.layout.GridPane;
+import javafx.beans.property.ReadOnlyObjectWrapper;
+import javafx.geometry.HPos;
+import javafx.geometry.Insets;
+import javafx.geometry.Orientation;
+import javafx.scene.control.*;
+import javafx.scene.layout.*;
+import javafx.util.Callback;
 
 import javax.annotation.Nullable;
 import java.util.UUID;
@@ -149,5 +152,178 @@ public class CompensationRequestDisplay {
 
     public int incrementAndGetGridRow() {
         return ++gridRow;
+    }
+
+    public GridPane createCompensationList(TableView<CompensationRequestListItem> tableView, String headerString) {
+        GridPane compensationList = new GridPane();
+
+        TableGroupHeadline header = new TableGroupHeadline(headerString);
+        GridPane.setMargin(header, new Insets(10, 0, 0, 0));
+        GridPane.setRowIndex(header, 0);
+        compensationList.getChildren().add(header);
+        header.setMinHeight(20);
+        header.setMaxHeight(20);
+
+        tableView.setColumnResizePolicy(TableView.CONSTRAINED_RESIZE_POLICY);
+        tableView.setPlaceholder(new AutoTooltipLabel(Res.get("table.placeholder.noData")));
+        tableView.setMinHeight(90);
+        GridPane.setRowIndex(tableView, 1);
+        GridPane.setColumnSpan(tableView, 1);
+        GridPane.setMargin(tableView, new Insets(5, 10, -10, 10));
+        GridPane.setVgrow(tableView, Priority.ALWAYS);
+        GridPane.setHgrow(tableView, Priority.ALWAYS);
+        compensationList.getChildren().add(tableView);
+
+        createColumns(tableView);
+
+        return compensationList;
+    }
+
+    public ScrollPane createCompensationRequestDisplay() {
+        ScrollPane scrollPane = new ScrollPane();
+        scrollPane.setFitToWidth(true);
+        scrollPane.setFitToHeight(true);
+        scrollPane.setMinHeight(100);
+
+        AnchorPane bottomAnchorPane = new AnchorPane();
+        scrollPane.setContent(bottomAnchorPane);
+
+        gridPane.setHgap(5);
+        gridPane.setVgap(5);
+        ColumnConstraints columnConstraints1 = new ColumnConstraints();
+        columnConstraints1.setHalignment(HPos.RIGHT);
+        columnConstraints1.setHgrow(Priority.SOMETIMES);
+        columnConstraints1.setMinWidth(140);
+        ColumnConstraints columnConstraints2 = new ColumnConstraints();
+        columnConstraints2.setHgrow(Priority.ALWAYS);
+        columnConstraints2.setMinWidth(300);
+
+        gridPane.getColumnConstraints().addAll(columnConstraints1, columnConstraints2);
+        AnchorPane.setBottomAnchor(gridPane, 20d);
+        AnchorPane.setRightAnchor(gridPane, 10d);
+        AnchorPane.setLeftAnchor(gridPane, 10d);
+        AnchorPane.setTopAnchor(gridPane, -20d);
+        bottomAnchorPane.getChildren().add(gridPane);
+
+        return scrollPane;
+    }
+
+    public SplitPane createCompensationRequestPane(TableView<CompensationRequestListItem> tableView, String headerString) {
+        SplitPane compensationRequestPane = new SplitPane();
+        compensationRequestPane.setOrientation(Orientation.VERTICAL);
+        compensationRequestPane.setDividerPositions(0.3, 0.7);
+        compensationRequestPane.setStyle("-fx-box-border: transparent;");
+
+        compensationRequestPane.getItems().add(createCompensationList(tableView, headerString));
+        compensationRequestPane.getItems().add(createCompensationRequestDisplay());
+        return compensationRequestPane;
+    }
+
+    private void createColumns(TableView<CompensationRequestListItem> tableView) {
+        TableColumn<CompensationRequestListItem, CompensationRequestListItem> dateColumn = new AutoTooltipTableColumn<CompensationRequestListItem, CompensationRequestListItem>(Res.get("shared.dateTime")) {
+            {
+                setMinWidth(190);
+                setMaxWidth(190);
+            }
+        };
+        dateColumn.setCellValueFactory((tradeStatistics) -> new ReadOnlyObjectWrapper<>(tradeStatistics.getValue()));
+        dateColumn.setCellFactory(
+                new Callback<TableColumn<CompensationRequestListItem, CompensationRequestListItem>, TableCell<CompensationRequestListItem,
+                        CompensationRequestListItem>>() {
+                    @Override
+                    public TableCell<CompensationRequestListItem, CompensationRequestListItem> call(
+                            TableColumn<CompensationRequestListItem, CompensationRequestListItem> column) {
+                        return new TableCell<CompensationRequestListItem, CompensationRequestListItem>() {
+                            @Override
+                            public void updateItem(final CompensationRequestListItem item, boolean empty) {
+                                super.updateItem(item, empty);
+                                if (item != null)
+                                    setText(bsqFormatter.formatDateTime(item.getCompensationRequest().getPayload().getCreationDate()));
+                                else
+                                    setText("");
+                            }
+                        };
+                    }
+                });
+        dateColumn.setComparator((o1, o2) -> o1.getCompensationRequest().getPayload().getCreationDate().compareTo(o2.getCompensationRequest().getPayload().getCreationDate()));
+        dateColumn.setSortType(TableColumn.SortType.DESCENDING);
+        tableView.getColumns().add(dateColumn);
+        tableView.getSortOrder().add(dateColumn);
+
+        TableColumn<CompensationRequestListItem, CompensationRequestListItem> nameColumn = new AutoTooltipTableColumn<>(Res.get("shared.name"));
+        nameColumn.setCellValueFactory((tradeStatistics) -> new ReadOnlyObjectWrapper<>(tradeStatistics.getValue()));
+        nameColumn.setCellFactory(
+                new Callback<TableColumn<CompensationRequestListItem, CompensationRequestListItem>, TableCell<CompensationRequestListItem,
+                        CompensationRequestListItem>>() {
+                    @Override
+                    public TableCell<CompensationRequestListItem, CompensationRequestListItem> call(
+                            TableColumn<CompensationRequestListItem, CompensationRequestListItem> column) {
+                        return new TableCell<CompensationRequestListItem, CompensationRequestListItem>() {
+                            @Override
+                            public void updateItem(final CompensationRequestListItem item, boolean empty) {
+                                super.updateItem(item, empty);
+                                if (item != null)
+                                    setText(item.getCompensationRequest().getPayload().getName());
+                                else
+                                    setText("");
+                            }
+                        };
+                    }
+                });
+        nameColumn.setComparator((o1, o2) -> o1.getCompensationRequest().getPayload().getName().compareTo(o2.getCompensationRequest().getPayload().getName()));
+        tableView.getColumns().add(nameColumn);
+
+        TableColumn<CompensationRequestListItem, CompensationRequestListItem> uidColumn = new AutoTooltipTableColumn<>(Res.get("shared.id"));
+        uidColumn.setCellValueFactory((tradeStatistics) -> new ReadOnlyObjectWrapper<>(tradeStatistics.getValue()));
+        uidColumn.setCellFactory(
+                new Callback<TableColumn<CompensationRequestListItem, CompensationRequestListItem>, TableCell<CompensationRequestListItem,
+                        CompensationRequestListItem>>() {
+                    @Override
+                    public TableCell<CompensationRequestListItem, CompensationRequestListItem> call(
+                            TableColumn<CompensationRequestListItem, CompensationRequestListItem> column) {
+                        return new TableCell<CompensationRequestListItem, CompensationRequestListItem>() {
+                            @Override
+                            public void updateItem(final CompensationRequestListItem item, boolean empty) {
+                                super.updateItem(item, empty);
+                                if (item != null)
+                                    setText(item.getCompensationRequest().getPayload().getUid());
+                                else
+                                    setText("");
+                            }
+                        };
+                    }
+                });
+        uidColumn.setComparator((o1, o2) -> o1.getCompensationRequest().getPayload().getUid().compareTo(o2.getCompensationRequest().getPayload().getUid()));
+        tableView.getColumns().add(uidColumn);
+
+        TableColumn<CompensationRequestListItem, CompensationRequestListItem> confidenceColumn = new TableColumn<>(Res.get("shared.confirmations"));
+        confidenceColumn.setMinWidth(130);
+        confidenceColumn.setMaxWidth(confidenceColumn.getMinWidth());
+
+        confidenceColumn.setCellValueFactory((item) -> new ReadOnlyObjectWrapper<>(item.getValue()));
+
+        confidenceColumn.setCellFactory(new Callback<TableColumn<CompensationRequestListItem, CompensationRequestListItem>,
+                TableCell<CompensationRequestListItem, CompensationRequestListItem>>() {
+
+            @Override
+            public TableCell<CompensationRequestListItem, CompensationRequestListItem> call(TableColumn<CompensationRequestListItem,
+                    CompensationRequestListItem> column) {
+                return new TableCell<CompensationRequestListItem, CompensationRequestListItem>() {
+
+                    @Override
+                    public void updateItem(final CompensationRequestListItem item, boolean empty) {
+                        super.updateItem(item, empty);
+
+                        if (item != null && !empty) {
+                            setGraphic(item.getTxConfidenceIndicator());
+                        } else {
+                            setGraphic(null);
+                        }
+                    }
+                };
+            }
+        });
+        confidenceColumn.setComparator((o1, o2) -> o1.getConfirmations().compareTo(o2.getConfirmations()));
+        tableView.getColumns().add(confidenceColumn);
     }
 }
